+++ src/App.jsx
@@ -8,7 +8,7 @@ import Profile from "./Modules/Profile/profile";
 import LoginPage from "./pages/login";
 import ForgotPassword from "./pages/forgotPassword";
 import AcademicPage from "./Modules/Academic/index";
-import MessPage from "./Modules/Mess/index";
+import MessPage from "./Modules/Mess/pages/index";
 import ValidateAuth from "./helper/validateauth";
 
 export default function App() {
